Move copy and remove from FileBasedSink to GcsUtil.
